feat: remove legacy auth routes
