+++ package.json
@@ -11,7 +11,6 @@
     "@tanstack/react-query": "^5.90.7",
     "@telicent-oss/ds": "3.0.0",
     "@triply/yasqe": "^4.2.28",
-    "axios": "^1.15.1",
     "classnames": "^2.3.1",
     "graphiql": "^3.7.1",
     "graphql": "^16.9.0",
@@ -74,7 +73,6 @@
     "@telicent-oss/ontologyservice": "^1.1.0",
     "ws": "8.17.1",
     "@triply/yasqe/**/form-data": "^3.0.4",
-    "axios/**/form-data": "^4.0.4",
     "qs": "^6.14.1",
     "minimatch": "^3.1.4",
     "lodash": "^4.18.0",
