+++ src/App.tsx
@@ -3,7 +3,7 @@ import { useLocation, Route, Routes } from 'react-router-dom';
 import { AuthRedirectUri, Callback } from '@telicent-oss/ds';
 import config from 'config/app-config';
 
-import { Sparql, UserFetch, UserInfo, ProtectedRoutes } from 'components';
+import { Sparql, UserFetch, ProtectedRoutes } from 'components';
 import { ErrorPage } from 'lib';
 import TelicentGraphiQL from './components/Graphiql/TelicentGraphiQL';
 
@@ -30,14 +30,11 @@ const App = () => {
         </Route>
       </Route>
       <Route path="/health" element={<h3>Hello I'm Telicent Query</h3>} />
-      <Route path="/user-info" element={<UserInfo />} />
       <Route path="/error" element={<ErrorPage />} />
-      {config.featureFlags?.FF_AUTH_V2 && (
-        <Route
-          path="/auth-redirect-uri"
-          element={<AuthRedirectUri config={config.AUTH_V2_CONFIG_WITH_LOGOUT} />}
-        />
-      )}
+      <Route
+        path="/auth-redirect-uri"
+        element={<AuthRedirectUri config={config.AUTH_V2_CONFIG_WITH_LOGOUT} />}
+      />
     </Routes>
   );
 };
