--- src/components/InactiveProfile/InactiveProfileNew.jsx
@@ -1,28 +1,0 @@
-import React from 'react';
-import { AppBar, Button, FlexBox, Text, useAuth } from '@telicent-oss/ds';
-import config from 'config/app-config';
-import { Box } from '@mui/material';
-
-const InactiveProfileNew = () => {
-  const { login } = useAuth();
-  return (
-    <>
-      <AppBar appName="query" apps={config.apps} isElevated />
-      <FlexBox direction="column" gap={2} sx={{ p: 4 }}>
-        <FlexBox>
-          <Text variant="body1">
-            Your profile is currently inactive. Please contact your system administrator. You cannot
-            continue until your profile is activated and attributes have been set.
-          </Text>
-        </FlexBox>
-        <Box>
-          <Button color="primary" variant="outlined" onClick={login}>
-            Retry
-          </Button>
-        </Box>
-      </FlexBox>
-    </>
-  );
-};
-
-export default InactiveProfileNew;
