+++ src/components/index.js
@@ -1,4 +1,3 @@
-export { default as UserInfo } from './InactiveProfile';
 export { default as ProtectedRoutes } from './ProtectedRoutes';
 export { default as Sparql } from './Sparql/DisplaySparql';
 export { default as UserFetch } from './UserFetch';
