+++ src/config/__tests__/app-config.test.ts
@@ -45,12 +45,10 @@ describe('app-config', () => {
   });
 
   it('builds config from env', () => {
-    const envOverrides = {
-      ACCESS_URL: 'https://ACCESS_URL.example.test',
+    const envOverrides: Env = {
       GRAPHQL_URL: 'https://GRAPHQL_URL.example.test',
       SPARQL_URL: 'https://SPARQL_URL.example.test',
       BETA: true,
-      featureFlags: { FF_AUTH_V2: true },
       AUTH_V2_CONFIG: authConfig,
     };
     const { default: config, getConfig } = loadConfig(envOverrides);
@@ -58,7 +56,6 @@ describe('app-config', () => {
     expect({ inputs: envOverrides, output: config }).toMatchInlineSnapshot(`
       {
         "inputs": {
-          "ACCESS_URL": "https://ACCESS_URL.example.test",
           "AUTH_V2_CONFIG": {
             "authServerUrl": "https://auth.example.test",
             "clientId": "client-123",
@@ -69,12 +66,8 @@ describe('app-config', () => {
           "BETA": true,
           "GRAPHQL_URL": "https://GRAPHQL_URL.example.test",
           "SPARQL_URL": "https://SPARQL_URL.example.test",
-          "featureFlags": {
-            "FF_AUTH_V2": true,
-          },
         },
         "output": {
-          "ACCESS_URL": "https://ACCESS_URL.example.test",
           "APP_CONFIG_JSON": {
             "app_name": "App Name",
             "app_name_snake_case": "app_name",
@@ -97,9 +90,6 @@ describe('app-config', () => {
           "MAP_TILER_TOKEN": undefined,
           "SPARQL_URL": "https://SPARQL_URL.example.test",
           "beta": true,
-          "featureFlags": {
-            "FF_AUTH_V2": true,
-          },
         },
       }
     `);
@@ -107,9 +97,8 @@ describe('app-config', () => {
     expect(getConfig()).toBe(config);
   });
 
-  it('handles missing feature flags and beta false', () => {
-    const envOverrides = {
-      ACCESS_URL: 'https://ACCESS_URL.example.test',
+  it('handles beta false', () => {
+    const envOverrides: Env = {
       GRAPHQL_URL: 'https://GRAPHQL_URL.example.test',
       SPARQL_URL: 'https://SPARQL_URL.example.test',
       BETA: false,
@@ -120,7 +109,6 @@ describe('app-config', () => {
     expect({ inputs: envOverrides, output: config }).toMatchInlineSnapshot(`
       {
         "inputs": {
-          "ACCESS_URL": "https://ACCESS_URL.example.test",
           "AUTH_V2_CONFIG": {
             "authServerUrl": "https://auth.example.test",
             "clientId": "client-123",
@@ -133,7 +121,6 @@ describe('app-config', () => {
           "SPARQL_URL": "https://SPARQL_URL.example.test",
         },
         "output": {
-          "ACCESS_URL": "https://ACCESS_URL.example.test",
           "APP_CONFIG_JSON": {
             "app_name": "App Name",
             "app_name_snake_case": "app_name",
@@ -156,9 +143,6 @@ describe('app-config', () => {
           "MAP_TILER_TOKEN": undefined,
           "SPARQL_URL": "https://SPARQL_URL.example.test",
           "beta": false,
-          "featureFlags": {
-            "FF_AUTH_V2": false,
-          },
         },
       }
     `);
